User configurable schemas (#8)
* Round trip unknown blocks

* Round trip unknown attributes of known blocks

* round trip unknown marks

* Add a SchemaAdapter indirection

* Fix various bugs in cross-schema editing

* Paramterise building the playground as a demo

Problem: The playground is growing new features to test more complex
parts of the logic of the editor. This is good but the playground is
currently also built and deployed and then run inside an iframe in the
blog post introducing rich text. For the blog post we want a simpler
demo.

Solution: Add a built time parameter which checks the NODE_ENV and if
we're in 'production', just shows the simple side by side demo.

* fix eslint complaints for react example