Move to appendTransaction based architecture
Problem: The current architecture which always replaces an incoming ProseMirror transaction with a new transaction breaks many ProseMirror plugins - including the history plugin.

Solution: Instead of replacing the incoming transaction, just apply it to the automerge document and then check that the transaction we would have produced matches the incoming transaction. If the transactions don't match then a) we append a change which fixes up the incoming transaction without breaking plugins and b) log an error because it means we've done something wron.