Refactor initialization to make it easier to get started
Problem: The initialization of this library requires a lot of
familiarity with the various moving parts. This makes it difficult to
get started. 

Solution: Add an init function which gives users the three things they
need: a schema, a prosemirror document, and a plugin. These things can
be passed straight to the Editor component without needing to understand
how they all fit together.