only depend on @automerge/automerge-slim
Problem: the library currently depends on @automerge/automerge and
@automerge/automerge-repo. By importing these names we force
initialization of the webassembly modules which implement automerge.
This makes the library unusable for people who are initializing the
webassembly modules themselves.

Solution: as per [1] we depend on @automerge/automerge/slim which defers
library initialization to the user. We also entirely remove the
dependency on automerge-repo, instead using an interface representing
the DocHandle.

[1]: https://automerge.org/docs/library_initialization/